changed DecimalField to FloatField

+++ nosleguma_darbs/deposits/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Deposit(models.Model):
     deposit = models.IntegerField()
     term = models.IntegerField()
-    rate = models.DecimalField(decimal_places=2, max_digits=10)
+    rate = models.FloatField()
 
     def interest(self):
         s = 0
